requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mock==5.1.0
packaging==23.1
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
